refactor: Friend 도메인 아키텍처 개선
- infrastructure 패키지 삭제

- repository 패키지 추가 및, FriendRepository 단일 jpa 구현체로 수정

- FriendSearchService 및, FriendSearchRepository 를 FriendService, FriendRepository로 통합

- FriendSearchRepository 명세를 FriendRepository에서 상속하여 단일 FriendRepository로 해당 기능 활용할 수 있도록 수정
