+++ src/main/java/gravit/code/friend/repository/FriendRepository.java
@@ -1,17 +1,17 @@
-package gravit.code.friend.infrastructure;
+package gravit.code.friend.repository;
 
 import gravit.code.friend.domain.Friend;
 import gravit.code.friend.dto.response.FollowerResponse;
 import gravit.code.friend.dto.response.FollowingResponse;
+import gravit.code.friend.repository.custom.FriendSearchRepository;
+import java.util.Optional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.Optional;
-
-public interface FriendJpaRepository extends JpaRepository<Friend, Long> {
+public interface FriendRepository extends JpaRepository<Friend, Long>, FriendSearchRepository {
     boolean existsByFollowerIdAndFolloweeId(
             long followerId,
             long followeeId
