+++ src/main/java/gravit/code/friend/repository/strategy/HandleSearchStrategy.java
@@ -1,7 +1,7 @@
-package gravit.code.friend.infrastructure.strategy;
+package gravit.code.friend.repository.strategy;
 
 import gravit.code.friend.dto.SearchPlan;
-import gravit.code.friend.infrastructure.sql.select.FriendsHandleSearchQuerySql;
+import gravit.code.friend.repository.sql.FriendsHandleSearchQuerySql;
 import gravit.code.friend.support.HandleNormalize;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
