+++ src/main/java/gravit/code/friend/repository/strategy/NicknameSearchStrategy.java
@@ -1,7 +1,7 @@
-package gravit.code.friend.infrastructure.strategy;
+package gravit.code.friend.repository.strategy;
 
 import gravit.code.friend.dto.SearchPlan;
-import gravit.code.friend.infrastructure.sql.select.FriendsNicknameSearchQuerySql;
+import gravit.code.friend.repository.sql.FriendsNicknameSearchQuerySql;
 import gravit.code.friend.support.NicknameNormalize;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
